[TEST] unfinished Event into Occasion refactor
