+++ src/containers/legend/legend_events/legend_event/legend_event_hook.tsx
@@ -1,4 +1,4 @@
-import { useCalendarContext } from './../../../calendar/calendar_context';
+import { useCalendarContext } from '../../../calendar/calendar_context';
 import { Occasion } from '../../../../utils/utils';
 import { convertDateToHoursMinutes } from '../../../../utils/date';
